Refactored the test data example for issue #34. Also updated the docs a little
